requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
Django==4.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
fonttools==4.55.3
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
whitenoise==6.8.2
zopfli==0.2.3.post1
